requirements.txt
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.32.3
Flask==3.1.0
numpy==2.2.6
pandas==2.3.3
pandas-ta==0.4.71b0
websocket-client==1.9.0
websockets==15.0.1
matplotlib==3.10.7
pytest==8.2.0
schwab-py==1.5.1

